+++ src/main/kotlin/onku/backend/global/auth/config/SecurityConfig.kt
@@ -43,7 +43,7 @@ class SecurityConfig(
         private val MANAGEMENT_ENDPOINT = arrayOf( // 경총
             "/api/v1/kupick/manage/**",
             "/api/v1/points/manage/**",
-            "/api/v1/attendance/scan"
+            "/api/v1/attendance/manage/**"
         )
 
 //        private val EXECUTIVE = arrayOf("") // 회장단
